google-oauth/requirements.txt
asgiref==3.3.4
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
google-auth==1.32.0
idna==2.10
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.1.0
python-decouple==3.4
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5
